refactor: emoji -> echo로 도메인명 변경
제가 맡지 않은 부분은 일단 충돌 방지를 위해 emoji 그대로 두었습니다.
