+++ client/src/features/echo/api/deleteEmoji.ts
@@ -1,5 +1,5 @@
 import { api } from '@/app/lib/api';
-import { EmojiResponse } from '../type/emoji';
+import { EmojiResponse } from '../type/echos';
 
 export const deleteEmoji = async (emojiId: number): Promise<EmojiResponse> => {
   const response = await api.delete(`/emojis/${emojiId}`);
