+++ client/src/features/echo/api/getEmojis.ts
@@ -1,5 +1,5 @@
 import { api } from '@/app/lib/api';
-import { EmojiResponse } from '../type/emoji';
+import { EmojiResponse } from '../type/echos';
 
 export const getEmojis = async (commentId: number): Promise<EmojiResponse> => {
   const response = await api.get(`/emojis/${commentId}`);
