+++ client/src/features/echo/api/sendEmoji.ts
@@ -1,5 +1,5 @@
 import { api } from '@/app/lib/api';
-import { EmojiRequest, EmojiResponse } from '@/features/emoji/type/emoji';
+import { EmojiRequest, EmojiResponse } from '@/features/echo/type/echos';
 
 export const sendEmoji = async (emoji: EmojiRequest): Promise<EmojiResponse> => {
   const response = await api.post('/emojis', emoji);
