+++ client/src/features/echo/const/echoType.ts
@@ -1,0 +1,5 @@
+export const ECHO_TYPE = {
+  THANKS: '마음 고마워요',
+  TOUCHED: '감동 받았어요',
+  COMFORTED: '위로가 됐어요',
+};
