+++ client/src/features/echo/ui/Emoji.tsx
