--- client/src/features/emoji/const/emojiType.ts
@@ -1,3 +1,0 @@
-export const EMOJI_TYPE = {
-  HEART: '❤️',
-};
